Fixed Postgres database crashing after auto reload
Error: Duplicate primary key error after modification of code due to auto reload. The application is trying to insert data into the database upon each reboot.
Solution: Utilized a count function that returns the number of entries that already exists in the Postgres database, if the count is greater than 0 then insertion is not required.